requirements.txt
Django==3.0.6
djangorestframework
psycopg2
django-crispy-forms
django-allauth
pillow
django-autoslug
django-debug-toolbar
django-author
django-cors-headers
coreapi
pyyaml